Split monolithic printer implementation into modular files and eliminate leaky abstraction.
Remove is_einsum() method from CodePrinter interface (leaky type-check)
Add pure virtual format_term(const Term&) to CodePrinter for polymorphic term formatting
Move TammPrinter and EinsumPrinter concrete classes to separate headers (tamm_printer.h, einsum_printer.h)
Create implementation files tamm_printer.cc and einsum_printer.cc with all backend-specific logic
Remove Term::einsum_str() method - formatting now delegated to EinsumPrinter
Update Term::str() to dispatch via printer->format_term(), eliminating conditional branches
Add forward declaration of Term in code_printer.h for format_term interface
This refactoring enforces strict separation between the AST (Term, Linkage) and output syntax backends. The CodePrinter hierarchy can now be extended with new backends without modifying existing code or introducing type checks.

All 46 tests pass (37 algebraic + 9 numerical).
