+++ CMakeLists.txt
@@ -83,6 +83,8 @@ pybind11_add_module(_pdaggerq
         pq_graph/src/fusion.cc
         pq_graph/src/graph_printing.cc
         pq_graph/src/code_printer.cc
+        pq_graph/src/tamm_printer.cc
+        pq_graph/src/einsum_printer.cc
         pq_graph/src/vertex_printing.cc
         pq_graph/src/dot_generator.cc
         pq_graph/src/timer.cc
